# modules/apps/data-engine/data-engine-rest-impl/rest-openapi.yaml
components:
    schemas:
        DataDefinition:
            description:
                https://www.schema.org/DataDefinition
            properties:
                availableLanguageIds:
                    items:
                        type: string
                    type: array
                classNameId:
                    format: int64
                    type: integer
                dataDefinitionFields:
                    items:
                        $ref: '#/components/schemas/DataDefinitionField'
                    type: array
                dataDefinitionKey:
                    type: string
                dataDefinitionRules:
                    items:
                        $ref: '#/components/schemas/DataDefinitionRule'
                    type: array
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                defaultLanguageId:
                    type: string
                description:
                    additionalProperties:
                        type: object
                    type: object
                id:
                    format: int64
                    type: integer
                instanceId:
                    format: int64
                    type: integer
                name:
                    additionalProperties:
                        type: object
                    type: object
                siteId:
                    format: int64
                    type: integer
                storageType:
                    type: string
                userId:
                    format: int64
                    type: integer
            type: object
        DataDefinitionField:
            properties:
                customProperties:
                    additionalProperties:
                        type: object
                    type: object
                defaultValue:
                    additionalProperties:
                        type: object
                    type: object
                fieldType:
                    type: string
                id:
                    format: int64
                    type: integer
                indexType:
                    default: "all"
                    enum: [all, keyword, none, text]
                    type: string
                indexable:
                    type: boolean
                label:
                    additionalProperties:
                        type: object
                    type: object
                localizable:
                    type: boolean
                name:
                    type: string
                nestedDataDefinitionFields:
                    items:
                        $ref: '#/components/schemas/DataDefinitionField'
                    type:
                        array
                readOnly:
                    type: boolean
                repeatable:
                    type: boolean
                required:
                    type: boolean
                showLabel:
                    type: boolean
                tip:
                    additionalProperties:
                        type: object
                    type: object
                visible:
                    type: boolean
            type: object
        DataDefinitionPermission:
            properties:
                addDataDefinition:
                    type: boolean
                definePermissions:
                    type: boolean
                delete:
                    type: boolean
                roleNames:
                    items:
                        type: string
                    type:
                        array
                update:
                    type: boolean
                view:
                    type: boolean
            type: object
        DataDefinitionRule:
            properties:
                dataDefinitionFieldNames:
                    items:
                        type: string
                    type: array
                dataDefinitionRuleParameters:
                    additionalProperties:
                        type: object
                    type: object
                name:
                    type: string
                ruleType:
                    type: string
            type: object
...
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Data Engine"
    version: v2.0
openapi: 3.0.1
paths:
...
    "/data-definitions/{dataDefinitionId}":
        delete:
            operationId: deleteDataDefinition
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataDefinition"]
        get:
            operationId: getDataDefinition
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description:
                        ""
            tags: ["DataDefinition"]
        put:
            operationId: putDataDefinition
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description:
                        ""
            tags: ["DataDefinition"]
    "/data-definitions/{dataDefinitionId}/data-definition-field-links":
        get:
            operationId: getDataDefinitionDataDefinitionFieldLinks
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: fieldName
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                type: string
                    description:
                        ""
            tags: ["DataDefinition"]
    "/data-definitions/{dataDefinitionId}/data-definition-permissions":
        post:
            operationId: postDataDefinitionDataDefinitionPermission
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: operation
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataDefinitionPermission"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataDefinitionPermission"
                description:
                    ""
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataDefinition"]
...
    "/instances/{instanceId}/data-definitions":
        get:
            operationId: getInstanceDataDefinitionsPage
            parameters:
                - in: path
                  name: instanceId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: classNameId
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataDefinition"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataDefinition"
                                type: array
                    description:
                        ""
            tags: ["DataDefinition"]
        post:
            operationId: postInstanceDataDefinition
            parameters:
                - in: path
                  name: instanceId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description:
                        ""
            tags: ["DataDefinition"]
    "/instances/{instanceId}/data-definitions/{dataDefinitionKey}":
        get:
            operationId: getInstanceDataDefinition
            parameters:
                - in: path
                  name: instanceId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: dataDefinitionKey
                  required: true
                  schema:
                      type: string
                - in: query
                  name: classNameId
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description:
                        ""
            tags: ["DataDefinition"]
    "/sites/{siteId}/data-definition-permissions":
        post:
            operationId: postSiteDataDefinitionPermission
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: operation
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataDefinitionPermission"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataDefinitionPermission"
            responses:
                204:
                    content:
                        application/json: {}
                    description:
                        ""
            tags: ["DataDefinition"]
...